ansible-playbooks/sentinel-bot/requirements.txt
# Discord
discord.py>=2.3.2

# Async HTTP
aiohttp>=3.9.1
quart>=0.19.4

# Database
aiosqlite>=0.19.0

# SSH
asyncssh>=2.14.1

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
